pkcs11: Move public domain pkcs11.h to separate file to simplify updates
Source:

https://github.com/latchset/pkcs11-headers/blob/main/public-domain/3.2/pkcs11.h
